Refactored React app structure
